refactor: split modules; add hspec/tasty bindings; drop Tasty test suite
