+++ minithesis.cabal
@@ -19,11 +19,24 @@ library
   import: warnings
   hs-source-dirs: src
   exposed-modules:
-      Minithesis
+      Minithesis,
+      Minithesis.Hspec,
+      Minithesis.Tasty
+  other-modules:
+      Minithesis.Cache,
+      Minithesis.Runner,
+      Minithesis.Strategy,
+      Minithesis.TestCase
   build-depends:
       base >=4.14 && <5,
+      bytestring,
+      containers,
+      directory,
+      filepath,
       random >=1.2 && <1.3,
-      containers
+      hspec >=2.10,
+      tasty >=1.4,
+      tasty-hunit >=0.10
 
   default-extensions:
       ScopedTypeVariables
@@ -42,18 +55,3 @@ test-suite minithesis-hspec
       hspec >=2.10,
       tasty,
       tasty-hspec
-
-test-suite minithesis-tasty
-  import: warnings
-  type: exitcode-stdio-1.0
-  hs-source-dirs: test
-  main-is: TastyMain.hs
-  other-modules:
-      Minithesis.Spec
-  build-depends:
-      base,
-      minithesis,
-      directory,
-      tasty >=1.4,
-      tasty-hspec >=1.1,
-      hspec >=2.10
